+++ back/src/domains/booking/controller/booking.controller.ts
@@ -27,10 +27,10 @@ import { Request } from 'express';
 import { USER_STATUS } from '../../../auth/const/userStatus.const';
 import { SessionAuthGuard } from '../../../auth/guard/session.guard';
 import { SeatStatus } from '../const/seatStatus.enum';
-import { BookingAmountReqDto } from '../dto/bookingAmountReqDto';
-import { BookingAmountResDto } from '../dto/bookingAmountResDto';
-import { BookReqDto } from '../dto/bookReqDto';
-import { BookResDto } from '../dto/bookResDto';
+import { BookingAmountReqDto } from '../dto/bookingAmountReq.dto';
+import { BookingAmountResDto } from '../dto/bookingAmountRes.dto';
+import { BookReqDto } from '../dto/bookReq.dto';
+import { BookResDto } from '../dto/bookRes.dto';
 import { SeatsSseDto } from '../dto/seatsSse.dto';
 import { ServerTimeDto } from '../dto/serverTime.dto';
 import { WaitingSseDto } from '../dto/waitingSse.dto';
