+++ back/src/domains/reservation/dto/reservationCreate.dto.ts
@@ -1,7 +1,7 @@
 import { ApiProperty } from '@nestjs/swagger';
 import { IsInt } from 'class-validator';
 
-import { ReservationSeatInfoDto } from './reservationSeatInfoDto';
+import { ReservationSeatInfoDto } from './reservationSeatInfo.dto';
 
 export class ReservationCreateDto {
   @ApiProperty({
