+++ back/src/domains/reservation/service/reservation.service.ts
@@ -11,10 +11,10 @@ import { Event } from '../../event/entity/event.entity';
 import { Place } from '../../place/entity/place.entity';
 import { Section } from '../../place/entity/section.entity';
 import { Program } from '../../program/entities/program.entity';
-import { ReservationCreateDto } from '../dto/reservationCreateDto';
-import { ReservationIdDto } from '../dto/reservationIdDto';
-import { ReservationSeatInfoDto } from '../dto/reservationSeatInfoDto';
-import { ReservationSpecificDto } from '../dto/reservationSepecificDto';
+import { ReservationCreateDto } from '../dto/reservationCreate.dto';
+import { ReservationIdDto } from '../dto/reservationId.dto';
+import { ReservationSeatInfoDto } from '../dto/reservationSeatInfo.dto';
+import { ReservationSpecificDto } from '../dto/reservationSepecific.dto';
 import { Reservation } from '../entity/reservation.entity';
 import { ReservedSeat } from '../entity/reservedSeat.entity';
 import { ReservationRepository } from '../repository/reservation.repository';
